requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
cssselect2==0.6.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.6
django-allauth==0.51.0
django-bower==5.2.0
django-cors-headers==3.13.0
django-filer==2.2.1
django-filter==22.1
django-import-export==2.8.0
django-jet-reboot==1.3.1
django-js-asset==2.0.0
django-libs==2.0.3
django-mptt==0.13.4
django-polymorphic==3.1.0
django-scheduler==0.9.6
django-tinymce==3.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
dropbox==11.32.0
easy-thumbnails==2.8.1
et-xmlfile==1.1.0
gunicorn==20.1.0
icalendar==4.1.0
idna==3.3
lxml==4.9.1
Markdown==3.3.7
MarkupPy==1.14
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.2.0
ply==3.11
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python-decouple==3.6
python-memcached==1.59
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
reportlab==3.6.11
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
stone==3.3.1
svglib==1.3.0
tablib==3.2.1
tinycss2==1.1.1
Unidecode==1.1.2
urllib3==1.26.10
webencodings==0.5.1
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
